pythonClientForChatGpt/requirements.txt
openai>=1.3.0
pdfplumber>=0.10.2
python-dotenv>=1.0.0
typing-extensions>=4.8.0